Create Get-EssentialsUsersWithAdmincount
List Entrust Identity Essentials users which have their account in AD configured with attribut AdminCount set to 1 (protected by AdminSDHolder). 
Users who has Admincount set to 1 and is a member of a protected group, will be shown in red. 
Users who has Admincount set to 1 and is not a member of a protected group, will be shown in green.
Users that can´t be found in AD, will be listed in yellow.